requirements/base.txt
Django==1.6.2
South==0.8.4
requests==2.2.1
django-tastypie==0.11.0
django-crispy-forms==1.2.3
